Initial creation of update process
This is the initial update for the toolbox update process.

I'm not sure if this should stay in the admin or be moved into the cPP-Toolboxes repo (at least once complete).

I do not know how much detail we want to get into for any specific scenario. I figure this is a good place to start.