certified biometrics
This is to close #31.

The idea here is to update the diagram to show that biometric sensors could be both trusted and untrusted. I have updated the diagram to show this.

I also edited the ESR/Assumptions about the biometrics. I removed the assumption (I'm not sure you can say biometrics are an ESR and then that they are an assumption, but I think that by specifying trusted biometrics we should be OK in setting it for an ESR to need to be certified).