CMFA published (#40)
* Initial update to ESR

This is the updated ESR with proposed changes for discussion

* Editorial changes

This is to close #17

I also found some misspellings.

Graphic was updated to add another alternative line from the biometrics to the CMFA PAD

* updates from call

Updates based on the 1/14 call

* name change

just a filename change

* Update CMFA_ESR.adoc

* Update CMFA_ESR.adoc

* bullets added

As a quick fix for display, I added bullets for these points. Without the bullets they don't get displayed (adoc treats them as markers).

We can edit this differently before release.

* BEW update

I have made several edits. The notes at the end have been reformatted as admonitions so they are clearly not part of the intended text of the final document.

* Multiple updates base on Issues

This is to close the following:

Close #30
Close #33
Close #32
Close #29

There are also some minor editorial fixes

* Revert "Multiple updates base on Issues"

This reverts commit cec36fea804576745b5e50b6dd39e08f8eb001ef.

* Updates based on issues

This is to close the following:

Close #30
Close #33
Close #32
Close #29

There are also some minor editorial fixes

* certified biometrics

This is to close #31.

The idea here is to update the diagram to show that biometric sensors could be both trusted and untrusted. I have updated the diagram to show this.

I also edited the ESR/Assumptions about the biometrics. I removed the assumption (I'm not sure you can say biometrics are an ESR and then that they are an assumption, but I think that by specifying trusted biometrics we should be OK in setting it for an ESR to need to be certified).

* input sensor trust

Adding an ESR to explicitly assign trust to input sensors. This is actually to close #22 by instead of worrying specifically about how to check any input, that the vendor must define how it is trusted.

* cleaning up

one word change and cleaning up biometric ESR images that aren't needed.

* update based on feedback

edit based on @gregott comment

* Update CMFA_ESR.adoc

edit based on 2/23 call to remove phrase "and protect the sensitive data during the locked state".

* Updated ESR

This is an attempt to be a "final" version of the ESR that could be sent for wider review.

* Revert "Updated ESR"

This reverts commit d8319bd745c43495ca21fa22b8d5aa39c1f6800b.

* Update for wider review

This is hopefully sufficient for wider review.

* removal of notes

to clear out missed internal notes

* Update CMFA_ESR.adoc

Fix typos and replace "verification" with "biometric verification" because term "verification" are used to express signal verification and biometric verification and it could confuse readers.

* re-add original ESR

I had accidentally deleted the original ESR (I wasn't thinking about how the merge to the main branch would work), so this adds the current version back so there won't be a conflict on the final approval.

* Update CMFA_ESR.adoc

Co-authored-by: Greg Fiumara <gregory.fiumara@nist.gov>

* Update BIO_ESR.adoc

Co-authored-by: Greg Fiumara <gregory.fiumara@nist.gov>

* Update CMFA_ESR.adoc

changes from the call

* Final update based on approval date

Changed version/date

Co-authored-by: n-kai <n-kai@ipa.go.jp>
Co-authored-by: Greg Fiumara <gregory.fiumara@nist.gov>